karyotype_analyzer/src/image_processor.py

import cv2
import numpy as np

def preprocess_image(image_path):
    # Load the image
    img = cv2.imread(image_path, cv2.IMREAD_GRAYSCALE)
    
    # Apply some basic preprocessing
    img = cv2.GaussianBlur(img, (5, 5), 0)
    _, img = cv2.threshold(img, 0, 255, cv2.THRESH_BINARY + cv2.THRESH_OTSU)
    
    return img

def detect_chromosomes(preprocessed_image):
    # Find contours in the image
    contours, _ = cv2.findContours(preprocessed_image, cv2.RETR_EXTERNAL, cv2.CHAIN_APPROX_SIMPLE)
    
    # Filter contours based on area to remove noise
    min_area = 100  # Adjust this value based on your images
    chromosomes = [cnt for cnt in contours if cv2.contourArea(cnt) > min_area]
    
    return chromosomes

def analyze_karyotype(image_path):
    preprocessed = preprocess_image(image_path)
    chromosomes = detect_chromosomes(preprocessed)
    
    # Count chromosomes
    chromosome_count = len(chromosomes)
    
    # Placeholder for more advanced analysis
    # You could add more sophisticated analysis here, such as chromosome classification
    
    return {
        "chromosome_count": chromosome_count,
        "analysis_result": "Normal" if chromosome_count == 46 else "Abnormal"
    }

if __name__ == "__main__":
    # Test the functions with a sample image
    test_image_path = "path_to_test_image.jpg"  # Replace with an actual test image path
    result = analyze_karyotype(test_image_path)
    print(f"Analysis result: {result}")
