karyotype_analyzer/requirements.txt
tensorflow
opencv-python-headless
pillow
numpy
